# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: The Jobs MV Phase 1 API
servers:
  - url: /api
paths:
...
  /jobs:
    get:
      operationId: listPublicJobs
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: experienceLevel
          in: query
          schema:
            type: string
            enum: [Entry Level, 1-2 Years, 3-5 Years, 5+ Years]
        - name: candidateScope
          in: query
          schema:
            type: string
            enum: [Local, International, Both]
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicJob"
  /jobs/{id}:
    get:
      operationId: getPublicJob
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Active public job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicJob"
        "404":
          $ref: "#/components/responses/Error"
...
  /companies/{id}:
    get:
      operationId: getPublicCompany
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Public company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicCompany"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    ClerkCookieAuth:
      type: apiKey
      in: cookie
      name: __session
    ClerkBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    Disclosure:
      type: string
      enum:
        - provided
        - not_disclosed
        - not_provided
...
    PublicJob:
      type: object
      required:
        - id
        - title
        - department
        - companyName
        - location
        - createdAt
        - verifiedEmployer
        - offerInformationCompleteness
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        companyName:
          type: string
        location:
          type: string
        experienceLevel:
          type:
            - string
            - "null"
          enum: [Entry Level, 1-2 Years, 3-5 Years, 5+ Years, null]
        candidateScope:
          type:
            - string
            - "null"
          enum: [Local, International, Both, null]
        description:
          type: string
        expiresAt:
          type:
            - string
            - "null"
          format: date-time
        applicationDeadline:
          type:
            - string
            - "null"
          format: date-time
        publishedAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
        source:
          type:
            - string
            - "null"
        originalSourceUrl:
          type:
            - string
            - "null"
        verificationStatus:
          type: string
        verifiedEmployer:
          type: boolean
        offerInformationCompleteness:
          type: integer
          minimum: 0
          maximum: 100
        salaryMin:
          type:
            - integer
            - "null"
        salaryMax:
          type:
            - integer
            - "null"
        salaryCurrency:
          type:
            - string
            - "null"
          enum: [USD, MVR, null]
        salaryPeriod:
          type:
            - string
            - "null"
          enum: [month, null]
        salaryDisclosure:
          $ref: "#/components/schemas/Disclosure"
        serviceCharge:
          type:
            - string
            - "null"
        serviceChargeDisclosure:
          $ref: "#/components/schemas/Disclosure"
        otherAllowances:
          type:
            - string
            - "null"
        otherAllowancesDisclosure:
          $ref: "#/components/schemas/Disclosure"
        overtime:
          type:
            - string
            - "null"
        overtimeDisclosure:
          $ref: "#/components/schemas/Disclosure"
        accommodation:
          type:
            - string
            - "null"
        accommodationDisclosure:
          $ref: "#/components/schemas/Disclosure"
        meals:
          type:
            - string
            - "null"
        mealsDisclosure:
          $ref: "#/components/schemas/Disclosure"
        healthInsurance:
          type:
            - string
            - "null"
        healthInsuranceDisclosure:
          $ref: "#/components/schemas/Disclosure"
        annualLeave:
          type:
            - string
            - "null"
        annualLeaveDisclosure:
          $ref: "#/components/schemas/Disclosure"
        airTicket:
          type:
            - string
            - "null"
        airTicketDisclosure:
          $ref: "#/components/schemas/Disclosure"
        workingHours:
          type:
            - string
            - "null"
        workingHoursDisclosure:
          $ref: "#/components/schemas/Disclosure"
        weeklyOff:
          type:
            - string
            - "null"
        weeklyOffDisclosure:
          $ref: "#/components/schemas/Disclosure"
        probation:
          type:
            - string
            - "null"
        probationDisclosure:
          $ref: "#/components/schemas/Disclosure"
        contractLength:
          type:
            - string
            - "null"
        contractLengthDisclosure:
          $ref: "#/components/schemas/Disclosure"
    PublicCompany:
      type: object
      required:
        - id
        - companyName
        - verified
        - createdAt
      properties:
        id:
          type: integer
        companyName:
          type: string
        logoUrl:
          type:
            - string
            - "null"
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
...